login funtion ready
