+++ views/pages/home.ejs
@@ -5,7 +5,7 @@
     <p>Trade your home-made dishes</p>
 
     <div class="start-options">
-        <a href="/user/login" class="btn">Login</a>
+        <a href="/login/" class="btn">Login</a>
         <a href="/user/register" class="btn">Sign up</a>
     </div>
