+++ views/pages/login.ejs
@@ -9,9 +9,6 @@
 
     <h1>Login</h1>
 
-
-
-
         <fieldset class="required">
             <legend for="email">Email</legend>
             <input type="email" name="email" id="email">
