+++ views/pages/register.ejs
@@ -63,7 +63,7 @@
         <input type="password" name="psw-repeat" placeholder="" id="psw-repeat">
     </fieldset>
 
-    <span>Have an acount? <a href="/user/login">Login now!</a></span>
+    <span>Have an acount? <a href="/login">Login now!</a></span>
     <input type="submit" class="btn" id="submit" value="Create your account">
 
 </form>
